move unstack_dates to utils
